Fix sidebar icon vertical position shift when collapsing from expanded to collapsed (icon-only) mode
